Decouple PostAggregator and FieldExpression
